fix misc

+++ user/test/test_page_alloc.c
@@ -22,10 +22,10 @@ static char tmpbuf[PAGE_SIZE];
  */
 static inline void _validate_zones_data(void)
 {
-  int zid;
+  unsigned int zid;
 
   for (zid = 0; zid < ARRAY_SIZE(zones); zid++) {
-    assert((int) zones[zid].id == zid);
+    assert(zones[zid].id == zid);
     assert(zones[zid].description);
     assert(zones[zid].start < zones[zid].end);
     assert(zones[zid].boot_freepages >= zones[zid].freepages_count);
